% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.availability.R
\name{check.availability}
\alias{check.availability}
\title{Check availability}
\usage{
check.availability(id, hmdb_prefix = "http://www.hmdb.ca/metabolites/")
}
\arguments{
\item{id}{HMDB ID}

\item{hmdb_prefix}{= "http://www.hmdb.ca/metabolites/"}
}
\value{
TRUE/FALSE
}
\description{
This function checks if a HMDB ID has been revoked
}
