% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.index.list.R
\name{create.index.list}
\alias{create.index.list}
\title{Create index list}
\usage{
create.index.list(path)
}
\arguments{
\item{path}{Path to the downloaded HMDB XML database}
}
\value{
A data.frame with indices
}
\description{
This Function is used to generate initial hmdb_indices data.frame
which can be used to match CAS, InChI, InChIKey, chemspider ID, chebi ID,
Pubchem ID, metlin ID, Name or SMILES to the HMDB ID using function
get.id.
}
\details{
Function expects local path to the downloaded XML file from HMDB.
}
