% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.ppp.R
\name{extract.ppp}
\alias{extract.ppp}
\title{Extract Predicted Physical Properties PPP}
\usage{
extract.ppp(entry)
}
\arguments{
\item{entry}{HMDB Entry which can be gotten by get.entry}
}
\value{
Predicted Physical Properties
}
\description{
This function extracts Predicted Physical Properties from HMDB Entry
}
