% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.entry.R
\name{get.entry}
\alias{get.entry}
\title{Get Entry}
\usage{
get.entry(
  id,
  prefix = "http://www.hmdb.ca/metabolites/",
  check_availability = T
)
}
\arguments{
\item{id}{HMDB ID by which the entry should be access through}

\item{prefix}{= "http://www.hmdb.ca/metabolites/"}

\item{check_availability}{= T Option whether it should be check if entry has been revoked}
}
\value{
A List of all available properties
}
\description{
This function downloads the XML Entry of a specific HMDB ID.
}
