% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.R
\name{get.id}
\alias{get.id}
\title{Get ID}
\usage{
get.id(data, hmdb_indices, name_data, name_hmdb_indices, verbose = T)
}
\arguments{
\item{data}{Data.frame with column HMDB where the indices are stored}

\item{hmdb_indices}{Data.frame created by create.index.list}

\item{name_data}{name of column from data by which should be matched}

\item{name_hmdb_indices}{name of column from hmdb_indices by which should be matched}
}
\value{
Data.frame data with added HMDB indices
}
\description{
This function adds missing HMDB IDs depending on a given matching criteria.
}
