% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.id.R
\name{prepare.id}
\alias{prepare.id}
\title{Prepare ID}
\usage{
prepare.id(id)
}
\arguments{
\item{id}{Number or HMDB ID}
}
\value{
HMDB ID of length 7
}
\description{
This Function will append any ID given to it until it has reached length 7
and afterwards combines it with "HMDB".
This allows the access to the right HMDB Hyperlink
}
