% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.id.R
\name{get.id}
\alias{get.id}
\title{Get ID}
\usage{
get.id(data, name_data, name_hmdb_indices, verbose = T)
}
\arguments{
\item{data}{data.frame with column HMDB where the indices are stored}

\item{name_data}{name of column from data by which should be matched}

\item{name_hmdb_indices}{name of column from hmdb_indices by which should be matched Can be on of the following:
\itemize{
\item "name"
\item "cas_registry_number"
\item "smiles"
\item "inchi"
\item "inchikey"
\item "chemspider_id"
\item "chebi_id"
\item "pubchem_compound_id"
\item "metlin_id"
\item "kegg_id"
\item "canonical.smiles"
}}

\item{hmdb_indices}{Data.frame created by create.index.list}
}
\value{
Data.frame data with added HMDB indices
}
\description{
This function adds missing HMDB IDs to a data frame depending on a matching criteria given by the user.
}
